Added some array examples and speed examples

+++ examples/float_vs_int/double_speed.cpp
@@ -1,7 +1,7 @@
 int main() {
-    float a = 0;
+    double a = 0;
     for(int i=0;i<1e7;i++) {
-        a += (float)i/10 * (float) i;
+        a += (double)i/10 * (double)i;
     }
     return 0;
 }
